Rename stack.py to retaddr.py
